LayoutTests/fast/selectors/has-complexity-traversal-count-expected.txt


PASS :has() child-change invalidation produces correct styles
PASS :has() child-change invalidation is not O(n^2)
